Move stuff from LoadedAvatar to SpawnedAvatar
to reduce load when getting all avatars at once
